[Ide] Fix problems reloading multi-target project in type system
When a project file is edited the project is reloaded. The type
system service was reloading the project info for each framweork
and calling the Roslyn workspace's OnProjectReloaded. If the
target framework has changed then this would fail since a new
project id was generated. New project ids are now added, instead
of being reloaded, and unused old project ids are removed.